refactor: 컨트롤러와의 결합도를 낮추기 위해 DTO 위치 수정
