+++ src/test/java/gift/controller/ProductControllerTest.java
@@ -10,8 +10,6 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import gift.controller.request.ProductRequest;
-import gift.controller.response.ProductResponse;
 import gift.service.ProductService;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
